Move executionStackToResumeContext into BrainRunner
- ResumeContext is now an internal implementation detail of BrainRunner
- BrainRunner.resume() takes machine + webhookResponse instead of resumeContext
- External consumers (like Cloudflare DO) no longer need to know about
  ResumeContext - they just pass the machine with replayed events
- Simplifies the public API surface: Brain, BrainRunner, StateMachine
- Add TODO-cleanup.md to track remaining refactoring tasks
